backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
drf-nested-routers==0.93.4
channels==4.0.0
channels-redis==4.1.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
redis==5.0.0
celery>=5.3.4
python-dotenv==1.0.0
Pillow==10.0.1
django-filter==23.2
drf-yasg==1.21.7
firebase-admin==6.2.0
